Replace test file to main file
